+++ tests/test_unit/conftest.py
@@ -14,6 +14,7 @@ def setting_environmental_vars():
         "SETTINGS__COGNITE__CLIENT_SECRET": "super-secret",
         "SETTINGS__POWEROPS__READ_DATASET": "read_from_this_dataset",
         "SETTINGS__POWEROPS__WRITE_DATASET": "write_to_this_dataset",
+        "SETTINGS__POWEROPS__MONITOR_DATASET": "monitoring_to_this_dataset",
         "SETTINGS__POWEROPS__cogshop_version": "987",
     }
     os.environ.update(setting_vars)